Add crt1 implementations.
The assembler code only prepares a minimal C environment (which means, on
powerpc it reserves the minimum stack necessary to call a C function, and on all
architectures it initializes the frame pointer chain), and then calls _start_c
with the original stack pointer as first argument.

_start_c then calls __start_libc() with that same pointer and also a pointer to
main(). Pointers to _init() and _fini() are not needed, since for one, I don't
support them, and for two, they would only come up in static linking, where libc
can access them. They don't come up in dynamic linking, because in that case,
initialization is a bit more difficult, and the address of those functions would
be exported in DT_INIT and DT_FINI anyway.