Add simple skeleton.
As stated before, __start_libc() will initialize the libc, then run the
constructors, then call main() and exit().

The functions to call the constructors and destructors are moved to their own
files, since they both only work in static linking, and I want to exclude them
in dynamic linking if I ever get there. I checked the order in glibc, and it
calls both arrays forwards, so I'll do that as well.

In dynamic linking this will have to be done differently, since there is going
to be more than one module.