Make test program compile.
I had made a simple program that just returns 0. Of course, it also did
not compile for lack of a pthread_arch.h and a syscalls.h. And it gave
exec stack warnings, so I added a compiler flag that marks all assembler
files appropriately.

The syscalls are the ones from the kernel, with all the "old" variants
removed, and all the "new" variants renamed. I also got rid of the "16",
"32", and "64" suffixes.

For the 32-bit architectures, I depend on time64 support in the kernel.
I will not build anything new now that will break in 2038 without
warning.