Fix __stdio_exit() linking.
Referring externally to a function that has weak aliases is wrong. The
problem is that you are not in control of which version gets linked in
in that case. Here, there is a strong version of __stdio_exit() in
ofl.c, but also a weak alias in exit.c. Therefore the call to
__stdio_exit() from __toread_needs_stdio_exit() can be satisfied by
either one, and I do not necessarily get the ofl.c code linked in.

Now I add a new symbol __stdio_exit_needed() that only exists in ofl.c,
and it is referenced from toread.c and towrite.c, thereby causing ofl.c
to always be linked in when they get linked in.