Add fwrite(), fputs(), and puts().
Adding this, I noticed a few bugs.

First of all, I changed the design of the stdio write function to only
return non-negative values. I can see that writing failed by it
returning short, which it only does for final failure, and this way
makes it easier to write algorithms like fwrite().

fwrite() I immediately delegate into an internal function __fwritex(),
which is like fwrite(), but without locking and without the
multiplication.

__fwritex() has to do different things based on the buffering mode. This
is because ISO C defines fwrite() in terms of repeated calls to fputc(),
and I don't want to have to speculate how many differences from that
spec might be visible to a program.

In so doing, I noticed some bugs:

1. __overflow() didn't handle line ending characters correctly. In line
   buffered files, they must never be buffered and always written.
2. stdout was unbuffered. It is supposed to be line buffered or fully
   buffered in Sirius.