Add allocator.
Also adding a locking function and whatever else might be needed.

C23 adds two functions in the hope that they might be useful, and maybe
I will use that later, but for now, they only wrap free().

Speaking of: Programs that avoid calling free don't need the full
allocator. They get a simple bump allocator (also called high-water-mark
allocator) that cannot free anything. This necessitates one layer of
indirection between malloc() and the actual allocator, to redirect the
flow at link time.

I should probably start to compile the lib with -ffunction-sections, to
be able to prune the unused stuff out of there.

I am also adding both generic and arch-specific versions of a
count-trailing-zeroes primitive. I will need it later anyway. In this
case here I need it to identify the least bin containing chunks.

I also need to start recording the page size. On x86, the page size is
fixed, but on other architectures, it is variable and given in an aux
header.