Add underflow function.
This is called when getc_unlocked() notices that the buffer is empty. It
then tries to refill the buffer. Attempting to read from the buffer is
probably futile, but it prevents an issue if the function ever were to
be called with something in there, so I leave it in.