Add assert_fail.
As per spec, we write a message to standard error, then abort the
process. abort() still needs implemented, but that will happen with the
rest of the signal stuff, since it is tied to that.

We need to explicitly flush stderr here, in case the user call
setvbuf(stderr) for some odd reason. Since aborting does not go through
the usual channels, we cannot expect __stdio_exit() to take care of it.