Add timespec_get().
That function of course merely wraps clock_gettime() (in a namespace
preserving way). clock_gettime() is a function implemented in the VDSO,
so I have to call it through a function pointer.

Since I call through a function pointer anyway, may as well build the
entire architecture on it, so we do the dispatch only once. My
implementation has no problem with Y2038, so the initialization is valid
for the entire process lifetime.