Add abort().
Sometimes you pull on a thread and it just keeps going until the entire
sweater is unravelled. In this case, abort() by its nature requires
raise(). Due to _Fork() being possible to call in signal handlers,
though, both raise() and abort() must block signals as part of their
actions.

abort() must never return, so to avoid problems if SIGABRT is ignored or
blocked, if the raise() didn't work, it must un-ignore and unblock that
signal. But that requires adding sigaction().

sigaction() must, on architectures defining SA_RESTORER, link to the
restorer. So that must also be defined. Also, the architecture with
__libc_sigaction() and __sigaction() is something I use to handle the
translation between struct sigaction and struct ksigaction.

For now, I have removed SA_RESTORER from powerpc64, because that
architecture currently does not have a signal trampoline and ignores
SA_RESTORER entirely. I will have to look up how Linux 2.6.0 does it,
but that can come later.