Add POSIX time parser.
It is not quite correct. The algorithm finds the DST transition times
for the given year and works with that. Unfortunately, if the time is
given as local time, then I cannot reliably know the correct year. So if
a DST change happens right at the change of year (or within the DST
difference) I may evaluate the wrong time zone. I don't really know how
to fix it.

I still need a TZ file parser, but I have now also added the code that
finds the file.