Add scaling functions.
Basic idea is to construct a floating point constant of the form 2^n and
multiply that with the input number. That way, I don't need to handle
any special values on input, since the FPU will do it for me.

I need up to two pre-scaling steps, because the total range of exponents
is a little larger than twice the maximum value. On downscale, I need to
ensure that n remains below -<type>_MANT_DIG, to prevent double rounding
issues. This way, if the final result is a subnormal number, the
intermediate results are not. And if the intermediate results are
subnormal, the final result is an underflow to 0.