Add first sqrt() draft.
Using the Goldschmidt algorithm. I still need to research its convergence
properties and create the data table. Also, the other precisions, but
that also requires understanding the convergence.

For this algo I need an expanding 64->128 bit multiplication, and I
added it to cpu.h, since I will likely need it for the other precisions,
too. Plus it can be made faster in each on the 64-bit architectures if
the CPU we compile against has an instruction for that. This is the case
for x86_64 (mulq) and aarch64 (umulhd). PowerPC64 has mulhdu in the
POWER9 profile, and at this point I have no support for runtime
selection of such things, so only compile-time selection is possible.