Improve timezone handling.
Now a timezone is not defined by its opposite anymore. __time_to_tz() no
longer needs to find something that maybe can't be found, and mktime()
also does a simpler thing now. It now attempts to automatically
determine the timezone always (which is necessary, since the tzinfo file
might contain more than two time zones) and corrects for the DST only if
the user explicitly requested an adjustment that mismatches the found
time.