Correct mktime() timezone output.
My god, this gets complicated fast. But yes, in the end POSIX requires
that mktime() set the fields of the argument to the same values
localtime() would set given the result. This means if a time far into
the summer is given with tm_isdst=0, it is adjusted by the DST offset
and then a time with tm_isdst=1 is returned.

I am stretching the definition of "DST offset" somewhat, as with tzinfo
files there may be more than one DST or non-DST zone defined. In this
case, I always look up the zone the given time would be in as a local
time, and if that zone disagrees with the wishes of the caller, I adjust
it to the "standard" time zone (meaning the one given by the TZ string
at the end of the file) and then adjust the output accordingly.