Add stat functions.
All functions are in the end implemented in fstatat(), so the
workarounds can be shared. The lengthy code for fstat() is just a
workaround for older kernel's inability to use fstat() on O_PATH FDs.
Slightly less old kernels could use fstatat(), but failing that, the
only choice is to stat() the procfd name.

Otherwise, I am introducing a struct kstat. On architectures providing
fstatat, it corresponds to struct stat64 or struct stat in the kernel.