Add dirent functions.
Most of these are straightforward. The definition of fdopendir() makes
it unattractive to implement opendir() with it, because you have to
check so many things. telldir() is a bit of a curve ball, because there
is no check for overflowing offsets, but there also is no way to error
out from this function.