Add ppc fenv.
Most of this is straightforward and actually the same between ppc and
ppc64. Difference is that in ppc64, hard-fp is not optional.

Some fenv functions can be implemented portably in terms of other fenv
functions, at least with the restrictions of this implementation in
place.

The reason FE_DFL_MODE and FE_DFL_ENV are invalid pointers is that this
enables me to just implement different defaults at any later point
should the need arise.