Add nftw().
I am implementing the option of failing when a path longer than PATH_MAX
is encountered. The alternative would be skipping that entry, but I
doubt that is a good thing. It would lead to the walk missing files that
actually are there.

I am not implementing FTW_CHDIR. musl isn't either, and it would make
everything more complicated, and I haven't found any users anywhere.