Add glob().
Basic idea for me was to glob each FS level separately. Starting from
the current directory (identified by the empty string) or the root, we
add each level by adding strings to the list of intermediate results.

If the next level has no unescaped glob characters, we only textually
append the unescaped version of the string to all the input strings. If
the next level does have glob characters, we open each input string as a
directory and fnmatch the directory contents. If they do match, we add a
new string to the output list.

Since any input string can turn into something between no and many
output strings, the output string list is kept entirely separate from
the input list.

For memory management, I maintain three lists and two pointers to lists.
The list "noalloc" has no allocated string vector and no allocated
strings in that vector, and it is used to contain the initial string set
(which is always length 1, and contains either "" or "/"), and the
pseudo-result if there was no match and GLOB_NOCHECK was set (which is
always length 1 and contains the unchanged pattern.