Add close() and posix_close().
The newest POSIX has a problem with Linux's version of close(), namely
that Linux returns EINTR but closes the FD anyway. POSIX gives me two
options for this case: To return success or to return EINPROGRESS. It
also adds posix_close() where the behavior is supposed to be selectable.

Linux does not offer a selection of behavior. I can only run the syscall
and deal with the situation on return. Therefore I have no choice but to
define POSIX_CLOSE_RESTART as 0. For the application, there is little
difference in information between returning EINPROGRESS and returning 0.
Both mean the FD is closed, and whatever else happens is out of your
hands. But applications may misinterpret returning an error, and so I
shall return 0.

All of this is bloody pointless, since EINTR returns from close() don't
usually happen. They are defined for tape drives only, which is not a
particularly common use case. Otherwise, maybe something can happen with
NFS. But that's it. POSIX patched around a pretty irrelevant part of the
standard, at least for modern everyday life.