Add setjmp().
Design taken from musl. setjmp() saves the non-volatile registers,
longjmp() restores them. sigsetjmp(..., 0) just tail calls setjmp(),
whereas sigsetjmp(..., 1) calls setjmp() as a subroutine, then calls a C
function I call __sigsetjmp_tail() with the same sigjmp_buf and the
return value from setjmp(). Since sigsetjmp() can't use any stack, it
stores its data in the unused part of the sigsets.

__sigsetjmp_tail() just executes the appropriate sigprocmask syscall to
save the sigmask if the return value is 0, or restore it if it is
nonzero.

siglongjmp() just calls longjmp().

On PowerPC64, the TOC pointer needs some special love. setjmp() treats
it like any other non-volatile register. longjmp() almost does the same,
but also stores it in the TOC pointer restore slot on stack before
jumping to the return call site, so that if it is returning to a remote
call, the TOC pointer restore does not end up clobbering r2.

For sigsetjmp() on PowerPC64, the situation is more complicated, as it
calls a C language function, which is invalid unless r2 is set to libc's
TOC pointer at that point. So the internal function of setjmp used there
takes the caller's r2 in r5 (and longjmp restores it in the right
places), but that means that longjmp clobbers libc's r2 that was there
when setjmp() was called, so sigsetjmp() has to stash that along with
the other values in the unused parts of the sigmask.