Add i386 fenv.
Going a non-traditional route by having the public fenv structure not be
the same as the thing the CPU manuals call "floating point environment".
It is not necessary to record all that stuff, and if need be, it can be
constructed.

The x87 lacks a single instruction that would mean we can ignore the
whole thing. If there was an FLDSW operation, allowing us to set the
status word to arbitrary things, we could just never interact with the
x87_fenv at all. But it doesn't exist, and the only way to set the
status word to arbitrary values is as part of an environment.

However, the i386 ABI also states that all function must be entered with
an empty x87 stack, so that means I know the only valid tag word ahead
of time. Therefore I basically never need the FSTENV instruction.

FLDENV, I am reliably told, is slower than LDMXCSR, so I try to avoid
that instruction as much as possible, and it is possible on machines
with SSE. For that reason, almost all functions do something different
for SSE machines than for those without SSE.