Block sigs while holding abort lock.
The abort lock is needed for abort(), which is a signal safe function.
Therefore, everything that takes the lock must block signals before
doing so, so that signal handlers can call abort() without deadlocking.

The implementation-internal signals don't call abort() and so don't need
to be blocked. I was briefly debating whether SIGCANCEL needed blocking,
but we don't execute a cancel point while holding the lock, so it is
going to be fine.