Add f*lockfile API.
ftrylockfile() takes care of initializing a pseudo-lock from
single-threaded mode and other settings. It also takes care of the
re-locking case. Therefore flockfile() doesn't have to do that.

flockfile() calls the internal __flockfile(). If returns 1, we need to
unlock the file at some point, which means this is the first time the
file was locked by us, so it has to be added to the locked file list. If
the internal function returns 0 instead, we already had the lock, so we
must increase the lock count.

The API currently does not allow me to handle an overflow in the lock
count. But I fail to see any reasonable way how a normal application
might create a situation where the lock count overflows.

All the explicitly locked files are collected in a list. When I come to
write the threading code, the thread exit code will go through that list
and kill all the locks, causing dead locks for any thread attempting to
lock those files again. This is to prevent a resource-reuse issue if a
thread exits with locks held and then a new thread is started with the
same TID later.

That is not necessary for the internally locked files, since those can
never escape locked into the wild. If they can, I have a bug that needs
fixing.