Add some more trivial stdio functions.
These are all very simple functions implemented in terms of other
functions.

For vasprintf(), there is a simple implementation and a good one. I
chose the simple implementation (calling vsnprintf() twice) for now.
That implementation is very simple and obviously correct. However, it
does the printf() work twice (actually 4 times, since printf_core() is
called twice in vfprintf()).

A better implementation might be to create a FILE that re-allocates the
string as needed. But that is harder to get right.

A side effect of the current implementation is that asprintf() doesn't
pull in free(), which a realloc() based version would do. So the
tradeoffs are not as clear.