Add socket API.
Most of this simply calls the relevant syscalls. All architectures have
gained the socket calls as normal syscalls, but some kernel versions
might not offer those on socketcall architectures. I still prefer the
normal syscalls, since they don't require spilling to memory.

This is no argument for i386, where the arguments start out in memory,
but I have a feeling GCC won't be able to use the stack arguments
in-place anyway. And so the only way to use those would be assembler,
but I don't want to use assembler for a micro-optimization.

In sockatmark(), I avoid calling ioctl(), since my version of ioctl() is
not necessarily POSIX compliant. Also, inlining the syscall is probably
faster, seeing as that function would have to decode variable arguments,
which often turns into a huge block of code.