Add if_nameindex() and if_freenameindex().
if_nameindex::if_name must be a char*, both because POSIX requires that
and because the final element is identified by it being 0. Therefore it
cannot just be a char[]. However, IF_NAMESIZE is so small that moving
the name into a different allocation would just be overkill, so I
allocate it all in the same structure.

This makes if_freenameindex() supremely simple.

I do not check for all possible conditions of bad input, because
distrusting the kernel makes no sense.