Add pthread_exit().
To implement the self-cleanup of detached threads, I either need an
assembler function that does the munmap and then the exit, or I just
create an inline function that invokes a C function on another stack.
a_stackinvoke() is that function.

Note that if I ever implement dynamic linking, a_stackinvoke() cannot be
used to transition to the executable's entry point. I will need another
CPU function for that.

The main pthread_exit() function is pretty simple. First of all,
cancellation is disabled, since I want to implement cancellation in
terms of calling pthread_exit(), and I don't want to think too much
about re-entrancy.

POSIX doesn't say if the cleanup buffers and TSD destruction should be
skipped in case we exit the process. I will just do it in this order,
because I can only detect the case of exiting the last thread after
taking the thread-list lock, and no application could should ever
execute with the thread-list lock held.