Add pthread_cancel().
Most of this follows the musl playbook, because Rich Felker makes
compelling arguments. Cancel points come in two flavors: System calls
and simple cancel points. The only example of the latter type I can
think of right now is pthread_testcancel(), which is not supposed to
block.

Cancel point system calls go through the function __syscall_cp(). If
cancellation is not linked in or disabled, that just does a normal
syscall.

But if cancellation is enabled, we go through a special function
__syscall_cp_asm(), which cancels if __pthread_self()->cancel is
nonzero, else it does a normal but instrumented system call.

Cancellation itself is implemented with a signal. This serves the
purpose of interrupting blocking syscalls, as well as immediately
cancelling threads in asynchronous cancellation. The only other things
the cancel signal does are to redirect a thread inside of the cancel
point to __cancel(), and to block the cancel signal in the parent
context, then resend it. This is done to re-evaluate things if the
interrupted context is itself a signal handler.

pthread_cancel(self) does not have to send the signal at all. This is
because pthread_cancel() is not async-cancel safe and not async-signal
safe, so it can only be called with disabled or deferred cancellation
active, from the main program and not a signal handler. So none of the
things the signal handler does are important to pthread_cancel() itself.