Restore sig mask before async cancel.
To ensure atomicity, the cancel handler is registered with all signals
blocked. But when proceeding to pthread_exit(), we must unblock at least
SIGSYNCCALL for correctness, so let's just restore the sigmask, with the
added mask for SIGCANCEL as part of acting on the cancellation. Which
only exists to prevent infinite recursion if someone keeps spamming
SIGCANCEL with async cancellation in effect.