Add a_inc() and a_dec() and fix __flockfile().
Oh boy, that was an oversight. I just wrote the thread list lock code
(upcoming commit), and noticed that I didn't have atomic increment and
decrement abstractions handy. But I should have had them, seeing as I
had used a TID lock pattern before for __flockfile().

So I looked it over and noticed that the wait count was missing. That is
of course bad. If thread A holds the lock and threads B and C are
waiting for it, then when A unlocks, it will wake up thread B, but then
when thread B unlocks, it will not know thread C is still waiting.

So the waiters count helps with that. But then I wondered why we have a
waiters bit. And I pondered and pondered and didn't find an answer. I
remember Rich Felker talking about it on the musl mailing list, but
couldn't find the message.

Eventually I saw the bug: The unlock code was wrong. It read the
waiters count after setting the lock to free. In case of the TLL, this
will be no problem, but for files, it is one. A huge one.

As soon as the lock word is set to 0, another thread might come in and
take the lock. That taking of the lock might be inside of fclose()
(which is missing the lock right now, but that is easily fixed), and
thus lead to eventually freeing the memory of the file structure.

But all this might happen while the original unlocker is suspended. Then
it comes back and tries to read the waiters count out of freed memory,
causing undefined behavior.

So __funlockfile() must read the waiters count before setting the lock
word. And that is what the waiters bit is for: The waiters count might
be stale by the time the atomic swap completes, but if a new thread
arrived at the lock between the unlocker reading the waiters count and
performing the atomic swap, then it definitely set the waiters bit,
and so we definitely know to call __futex_wake in this case.