Add lock to fclose().
It is possible to call fclose() with a file still locked, but when it
becomes unlocked, we can free it. For that reason, we must take the lock
on the file, and we will not release it because in the main case, the
file will be freed.

For the three standard streams, they will just remain locked forever in
this way. I think that is OK, because after fclose(), they are closed,
and further references to them are undefined behavior.