Add hash table.
Simple open-addressing hash table with quadratic probing. The probe
sequence is the triangle numbers, and the hash table always has a size
of a power of two. This way, the triangle numbers will always eventually
enumerate all hash table elements.

Putting it all in the same file, since they all need the same
definitions and the same data.