Fix pthread_kill and pthread_cancel.
This is a complicated interaction. First of all, pthread_kill(). That
one is supposed to be async-signal-safe, so if we don't block signals,
we might end up in a deadlock situation.

Unfortunately, to avoid a deadlock we actually have to block signals
everywhere we take the killlock. Hence the change to pthread_cancel().
Might as well just call pthread_kill(), which already does everything
that is needed.

But pthread_cancel() is supposed to be async-cancel safe. If the thread
gets asynchonously cancelled while holding the killock, then
pthread_exit() will deadlock. To avoid this, we block cancellation by
blocking all signals.