Make quick_exit() signal-safe.
at_quick_exit() and quick_exit() are defined to be thread-safe in POSIX,
and therefore must synchronize on the use of the global list of
destructors. But quick_exit() must also be signal-safe, and in
particular be able to be called while at_quick_exit() is ongoing
elsewhere. Calling quick_exit() more than once is explicitly labelled
undefined behavior, so I am not worried about re-entrancy. But the lock
quick_exit() takes must not be held while signals are unlocked. And for
that reason, I block signals in at_quick_exit().

I technically don't need to block signals in quick_exit(), but it is
good form to observe the same limitation in all places where a lock is
taken.