Add handler set.
For posix_spawn(), I need a set of all signals that might have a
handler. It is only needed as an optimization, so I don't have to query
the signal handler for all of the signals.

To avoid a lock, I add a new atomic to OR in the new bits as needed.
Reading can be done normally; in the worst case I am getting too many
bits, which is a case I already have to handle.