Add __timedwait().
I am adding __timedwait_cp() along with it, since I will likely need a
cancel point version of this later on.

The clockid must specify a real time clock, not a CPU time clock. This
allows me to use the APIs that measure time against the real time clock
rather than the monotonic. Not that it matters much since the timeout
given to the syscall is relative.

There is a historic bug in Linux where even non-interrupting signals (so
signals with handlers with SA_RESTART set) would cause SYS_futex to
return -EINTR. For this reason, I am now tracking whether any such
signal handlers are installed, and retry on EINTR if not.