Add faccessat().
I need it to implement execvpe(), and therefore I do provide emulation
on old kernels. In theory I could drop it on 32-bit archs, since the
baseline there is newer than SYS_faccessat2 existing, but on 64-bit
archs that is not the case. And since I need the code anyway, might as
well compile it all the time.