Add fnmatch().
Complicated function made worse by POSIX overgeneralizing. I have not
implemented collation elements correctly, because they break an
important invariant in the code. And nobody uses them anyway, since they
are so badly specified.

For now, collation elements are just skipped as unsupported, but
unfortunately, they are a valid start or end point of a range in POSIX,
and that is not implemented here. Maybe I'll look into how other libcs
are doing that later.

Anyway, the main algorithm is a recursive backtracking solver. I know,
recursion is not nice, and O(n) stack frames (in the length of the
pattern) can blow up badly. Maybe I'll look into a better one later. I
don't want to copy musl outright, but they are the only place mentioning
this "Sea of stars" algorithm, which I have so far failed to find
descriptions of since unfortunately "Sea of stars" is also the name of a
video game.