Add exp().
Not the other ones in the same family yet, I still need to do more
research for those. But this does the same thing as the fdlibm version,
only polished up a lot and replacing integer compares with FP compares.

I added FORCE_EVAL() to make the compiler emit seemingly useless code
just for the side effect of causing exceptions. But this also added
useless warnings all over the place.