Add wctype functions.
Oof, this was a lot of work. I am adding the Unicode text files I used
as base for everything else, and the intermediate files I created. In
the end I could make do with only three large sets (alpha, printable,
and the case map), and that should deal with these functions for a
while.