Pass on resolution errors.
If process_name() fails, then result() contains some partial result that
doesn't fit the glob pattern given, at least not necessarily (it would
if the error occurred when processing the last name of the pattern). It
is just way easier to release the memory and pass on the result than to
identify if we were in the last place first. So let's just do that.