Add atan2() and atan().
atan() basically works by approximating the arctangent with a Taylor
series. The series is bloody long (11th degree), and that still only
converges for small x. So for larger x, we use reflection formulas to
get the argument in range. Strangely, I cannot find these formulas
anywhere on the net, but fdlibm does it this way.

atan2() is then a wrapper around atan() which handles more corner cases.

fdlibm tries to avoid the division at the end of atan2() if the two
arguments are too far apart, and I have no idea why. atan() can handle
overflow and underflow itself, so no need to do that in the wrapper. For
this reason, I have not copied that part.