Add double trig functions.
The heart of the algorithm, __rem_pio2, is still missing, but it will
come soon.

The division of labor among these functions is quite straightforward:
The public functions take care of exceptional arguments (none of these
functions has a limit in infinity, so we always return NaN in those
cases), as well as argument reduction (and the case where argument
reduction isn't needed). The kernel functions then calculate the main
functions on [-π/4; π/4].