Add getcwd().
And with this, it really becomes time to curate my warning options,
because GCC spuriously gives a warning here. The function itself is
complicated by the kernel trying to be helpful. I'm now doing the same
thing as musl, rejecting "(unreachable)" paths with ENOENT. That way at
least there's some consistency.