Add __rem_pio2().
Purpose of the function is to calculate n = round(x*2/π), and then also
x - n π/2. Musl and fdlibm have an ungodly complicated function here
that does something I cannot understand, no matter how hard I try. Parts
of it make sense after reading the Shewchuk paper, but other parts still
don't. I also think the effort they go to is severely misplaced here,
which I'll justify further down.

I don't go that far. I have 53 bits of the constant π/2 saved, then use
the formulas from Shewchuk's paper to calculate -n*π/2 with two words of
precision, because the multiplication already expands things. Then I
only need to add x to the result and get my desired output to two words
of precision.

My interpretation of floating-point numbers is that they are a stand-in
for an interval of real numbers, namely all those real numbers closer to
the representative than all representable numbers. There are a few
caveats to this definition, but mostly this means the interval is the
given number ±1/2 ULP. And the ULP is always a power of two in a binary
FP system.

If the ULP is 8 or greater, then it is larger than 2π, and therefore,
there is always a multiple of 2π in the interval, and we can just return
0. If the ULP is 2, then it is still larger than π/2, and therefore I
can give the remainder as 0. The divisor I have to calculate, though.

For all others, I perform the normal algorithm. Shewchuk's paper says
that the tuple (p.lo, q.lo, q.hi) will end up in *ascending* order,
except that any of them may be 0. So I have to take care to actually
filter out the zeroes.