Add __setxid() and __synccall().
These are necessary to run syscalls from every thread of a process. No
process should ever use multi-threaded credential changes, of course,
but I still need to implement these correctly.

In case of __setxid(), partial success will be answered with an
immediate abort(). The process is in an untennable state then and can
only be ended immediately.