Add fork(), _Fork(), and pthread_atfork().
These functions are somewhat linked. I will later change fork() to take all
implementation-internal locks. _Fork() is supposed to be async-signal-safe, so
it cannot do this.

pthread_atfork() is like atexit(), but I do need to be able to iterate in both
directions. That's why it gets a doubly linked list. I chose 32 registrations
per block for symmetry with atexit(), but POSIX is silent on the issue.