Add locale support.
It was already rudimentary, but we could not yet switch to UTF-8 mode. Now we
can with both setlocale() and newlocale().

I'm holding off on getlocalename_l() for now, since that looks to be bit of a
pain. The others are reasonably straightforward. musl takes an exclusive lock in
newlocale(), though, which I need to research what that's for.