Fix memory management in glob().
First of all, at the start of the function I declare a pointer "p" that only
lives for the resolution block. However, if the pattern is just "/", that
pointer is still used in noalloc.paths[0], which would become dangling. For that
reason I need to move the declaration up to function level.

Second change: The "if (rv) return rv;" block after the resolution loop must not
be used for rv == GLOB_NOMATCH, because in that case I still need to handle the
GLOB_NOCHECK flag.

For globfree(), I need all result strings to be allocated, lest I cause
undefined behavior. But for the case of pat just being "/" and the no-match case
with GLOB_NOCHECK, the result is a single non-allocated string. To solve this,
if the result is in noalloc at the end of the function, I create an allocated
copy. I could conceivably get rid of this copy for the pat=="/" case by creating
a global named constant for the root path (because then globfree() can ensure
that a pointer isn't a pointr to that object), but that really doesn't help for
the GLOB_NOCHECK case. So I need most of this code either way.