glob(): Limit glob character search to current file name.
In the unescaped case, my has_unescaped_globchars() function would search for
glob characters in the entire string, when it was supposed to be limited by the
length given. There is unfortunately no length-limited version of strcspn(), so
now I am still doing that, but returning 0 if the found characters is beyond the
end of the current file name.