Fix constraint violation in syscall.h
I was test compiling with stricter warning settings for once, and the
output was very noisy, but occasionally a good thing is reported. In
this case, if syscall was instantiated with 0 arguments, then after
finitely many steps __SYSCALL_NARGS2 is instantiated with nothing in the
variadic arg, which is not allowed in C99, but GCC allows it.