Remove ambiguous casefolds.
Unicode sometimes has weird ideas about casefolding. But I do think it should
round-trip. For that reason, the mapping must be injective. And so I have
removed all casemap rows that led to an already-used character.

Conflict resolution strategy was for me to look at the characters in question
and decide on which is the "best" by gut feeling. For example, there was a
mapping from the combining iota subscript to the non-combining iota. I removed
that, because case mapping should not turn combiners into non-combiners and vice
versa.

This means that I now consider the latin long s an uncased character. Oh well, I
gotta disappoint some people no matter what I do here.