Fix bug in __run_atexit_funcs().
Of course, when reverting to an earlier block, I must also reset the number of
pointers in the current block back to 32. And of course I copied the same
mistake into __run_quick_exit_funcs(), because I don't always make mistakes, but
when I do, I make sure to copy them.