Add __lookup_ipliteral().
That function is a helper needed in two places for __lookup_name(). It accepts
either IPv4 addresses or IPv6 addresses, optionally augmented with a scope ID in
the form of an interface index or name. Those only make sense as direct input or
from /etc/hosts, but not from the network.