Add __dns_create_query() and __dns_process().
Both of these derive from the same RFC, and are mostly about DNS packets by
themselves. Note that this is a stub resolver, and not meant to counter adverse
attacks. Currently, __dns_process() does not check if the result matches the
query. This may be the biggest weakness for now.