Add __dns_transact().
This should do the DNS transaction, including, if necessary, the TCP
stuff with the TCP fastopen if supported.

This is how huge functions happen: You just start writing and never
stop.

I am using a little-known feature of poll(): If a given pollfd has an fd
member < 0, then it is ignored for the main poll. This saves me from
adding another permutation to put all the active connections at the
start of the pollfd array.

I am using the older TCP fastopen definition, which was added in Linux
3.7. If an older Linux is used, the sendmsg() call will fail with EPIPE
instead, due to an attempt to send on an unconnected socket. That is why
I also add MSG_NOSIGNAL in the call, and fall back to non-blocking
connect.