Add dynamic lib.
Ooh boy, this is dumb. So I needed to add detection for a PIC-by-default
compiler, and noticed that mine was. However, it was actually a PIE-by-default
compiler and so the object files cannot be used to build a shared lib.

For now I am linking without "-nostdlib". If I ever change that, I will have to
re-add libgcc or whatever the compiler is using.

libc doesn't have to support internal interposition, and so I don't. But I still
have to allow interposition of data objects. For now, I have listed all exported
data objects in dynamic.list and am using that. This makes it easy to add
functions that might be interposed later.