glob: Fix memory errors.
There were a few dumb mistakes I found on review. Putting a '/' between a
candidate and a path should only happen when the candidate does not already end
in a '/', which unfortunately is the case if the glob pattern is an absolute
path.

When pruning non-existant entries at the end, we have to free the paths we are
overwriting. We also have to free the result path vector at the end. That is
definitely dead memory, even if the strings themselves survive.

And the final path vector allocation was short by a factor of one pointer size.